+++ apps/omg_api/lib/application.ex
@@ -19,68 +19,9 @@ defmodule OMG.API.Application do
   """
 
   use Application
-  use OMG.API.LoggerExt
-
-  def coordinator_setup do
-    deposit_finality_margin = Application.fetch_env!(:omg_api, :deposit_finality_margin)
-
-    %{
-      depositor: [finality_margin: deposit_finality_margin],
-      exiter: [waits_for: :depositor, finality_margin: 0],
-      in_flight_exit: [waits_for: :depositor, finality_margin: 0],
-      piggyback: [waits_for: :in_flight_exit, finality_margin: 0]
-    }
-  end
+  alias OMG.API.Supervisor
 
   def start(_type, _args) do
-    DeferredConfig.populate(:omg_api)
-
-    children = [
-      {OMG.API.State, []},
-      {OMG.API.BlockQueue.Server, []},
-      {OMG.API.FreshBlocks, []},
-      {OMG.API.FeeServer, []},
-      {OMG.API.RootChainCoordinator, coordinator_setup()},
-      OMG.API.EthereumEventListener.prepare_child(
-        service_name: :depositor,
-        synced_height_update_key: :last_depositor_eth_height,
-        get_events_callback: &OMG.Eth.RootChain.get_deposits/2,
-        process_events_callback: &OMG.API.State.deposit/1
-      ),
-      OMG.API.EthereumEventListener.prepare_child(
-        service_name: :in_flight_exit,
-        synced_height_update_key: :last_in_flight_exit_eth_height,
-        get_events_callback: &OMG.Eth.RootChain.get_in_flight_exit_starts/2,
-        process_events_callback: &exit_and_ignore_events_and_validities/1
-      ),
-      OMG.API.EthereumEventListener.prepare_child(
-        service_name: :piggyback,
-        synced_height_update_key: :last_piggyback_exit_eth_height,
-        get_events_callback: &OMG.Eth.RootChain.get_piggybacks/2,
-        process_events_callback: &exit_and_ignore_events_and_validities/1
-      ),
-      OMG.API.EthereumEventListener.prepare_child(
-        service_name: :exiter,
-        synced_height_update_key: :last_exiter_eth_height,
-        get_events_callback: &OMG.Eth.RootChain.get_standard_exits/2,
-        process_events_callback: fn exits ->
-          exits
-          |> Enum.map(&OMG.Eth.RootChain.get_standard_exit_utxo_pos/1)
-          |> exit_and_ignore_events_and_validities()
-        end
-      ),
-      {OMG.RPC.Web.Endpoint, []}
-    ]
-
-    opts = [strategy: :one_for_one, name: __MODULE__, max_restarts: 1000, max_seconds: 60_000]
-
-    _ = Logger.info("Starting #{inspect(__MODULE__)}")
-    :ok = :error_logger.add_report_handler(Sentry.Logger)
-    Supervisor.start_link(children, opts)
-  end
-
-  defp exit_and_ignore_events_and_validities(exits) do
-    {status, _events, db_updates, _validities} = OMG.API.State.exit_utxos(exits)
-    {status, db_updates}
+    Supervisor.start_link()
   end
 end
