+++ apps/omg_watcher/lib/block_getter/supervisor.ex
@@ -37,7 +37,7 @@ defmodule OMG.Watcher.BlockGetter.Supervisor do
       }
     ]
 
-    opts = [strategy: :one_for_all, name: __MODULE__, max_restarts: 1000, max_seconds: 60_000]
+    opts = [strategy: :one_for_all, max_restarts: 1000, max_seconds: 60_000]
 
     _ = Logger.info("Starting #{inspect(__MODULE__)}")
     Supervisor.init(children, opts)
